move client request handler into core::http1::client
